requirements.txt
Flask
yt-dlp
python-dotenv 